requirements.txt
awesome-slugify==1.6.5
colored==1.4.2
PyMySQL==0.9.3
regex==2020.1.8
Unidecode==0.4.21



